Split execution helper function tests out
