+++ execution_test.go
@@ -1667,86 +1667,6 @@ func TestQueryWithBoundaryFieldsAndNullsAboveInsertionPoint(t *testing.T) {
 	f.checkSuccess(t)
 }
 
-func TestExtractBoundaryIDs(t *testing.T) {
-	dataJSON := `{
-		"gizmos": [
-			{
-				"_bramble_id": "1",
-				"name": "Gizmo 1",
-				"owner": {
-					"_bramble_id": "1"
-				}
-			},
-			{
-				"_bramble_id": "2",
-				"name": "Gizmo 2",
-				"owner": {
-					"_bramble_id": "1"
-				}
-			},
-			{
-				"_bramble_id": "3",
-				"name": "Gizmo 3",
-				"owner": {
-					"_bramble_id": "2"
-				}
-			},
-			{
-				"_bramble_id": "4",
-				"name": "Gizmo 4",
-				"owner": {
-					"_bramble_id": "5"
-				}
-			}
-		]
-	}`
-	data := map[string]interface{}{}
-	expected := []string{"1", "1", "2", "5"}
-	insertionPoint := []string{"gizmos", "owner"}
-	require.NoError(t, json.Unmarshal([]byte(dataJSON), &data))
-	result, err := extractBoundaryIDs(data, insertionPoint)
-	require.NoError(t, err)
-	require.Equal(t, expected, result)
-}
-
-func TestTrimInsertionPointForNestedBoundaryQuery(t *testing.T) {
-	dataJSON := `[
-			{
-				"id": "1",
-				"name": "Gizmo 1",
-				"owner": {
-					"_bramble_id": "1"
-				}
-			},
-			{
-				"id": "2",
-				"name": "Gizmo 2",
-				"owner": {
-					"id": "1"
-				}
-			},
-			{
-				"id": "3",
-				"name": "Gizmo 3",
-				"owner": {
-					"_bramble_id": "2"
-				}
-			},
-			{
-				"id": "4",
-				"name": "Gizmo 4",
-				"owner": {
-					"id": "5"
-				}
-			}
-		]`
-	insertionPoint := []string{"namespace", "gizmos", "owner"}
-	expected := []string{"owner"}
-	result, err := trimInsertionPointForNestedBoundaryStep(jsonToInterfaceSlice(dataJSON), insertionPoint)
-	require.NoError(t, err)
-	require.Equal(t, expected, result)
-}
-
 func TestNestingNullableBoundaryTypes(t *testing.T) {
 	t.Run("nested boundary types are all null", func(t *testing.T) {
 		f := &queryExecutionFixture{
@@ -1978,1126 +1898,6 @@ func TestNestingNullableBoundaryTypes(t *testing.T) {
 		f.checkSuccess(t)
 	})
 
-}
-
-func TestBuildBoundaryQueryDocuments(t *testing.T) {
-	ddl := `
-		type Gizmo {
-			id: ID!
-			color: String!
-			owner: Owner
-		}
-
-		type Owner {
-			id: ID!
-			name: String!
-		}
-
-		type Query {
-			gizmos: [Gizmo!]!
-			getOwners(ids: [ID!]!): [Owner!]!
-		}
-	`
-	schema := gqlparser.MustLoadSchema(&ast.Source{Name: "fixture", Input: ddl})
-	boundaryField := BoundaryField{Field: "getOwners", Argument: "ids", Array: true}
-	ids := []string{"1", "2", "3"}
-	selectionSet := []ast.Selection{
-		&ast.Field{
-			Alias:            "_bramble_id",
-			Name:             "id",
-			Definition:       schema.Types["Owner"].Fields.ForName("id"),
-			ObjectDefinition: schema.Types["Owner"],
-		},
-		&ast.Field{
-			Alias:            "name",
-			Name:             "name",
-			Definition:       schema.Types["Owner"].Fields.ForName("name"),
-			ObjectDefinition: schema.Types["Owner"],
-		},
-	}
-	step := &QueryPlanStep{
-		ServiceURL:     "http://example.com:8080",
-		ServiceName:    "test",
-		ParentType:     "Gizmo",
-		SelectionSet:   selectionSet,
-		InsertionPoint: []string{"gizmos", "owner"},
-		Then:           nil,
-	}
-	expected := []string{`{ _result: getOwners(ids: ["1", "2", "3"]) { _bramble_id: id name } }`}
-	ctx := testContextWithoutVariables(nil)
-	docs, err := buildBoundaryQueryDocuments(ctx, schema, step, ids, boundaryField, 1)
-	require.NoError(t, err)
-	require.Equal(t, expected, docs)
-}
-
-func TestBuildNonArrayBoundaryQueryDocuments(t *testing.T) {
-	ddl := `
-		type Gizmo {
-			id: ID!
-			color: String!
-			owner: Owner
-		}
-
-		type Owner {
-			id: ID!
-			name: String!
-		}
-
-		type Query {
-			gizmos: [Gizmo!]!
-			getOwner(id: ID!): Owner!
-		}
-	`
-	schema := gqlparser.MustLoadSchema(&ast.Source{Name: "fixture", Input: ddl})
-	boundaryField := BoundaryField{Field: "getOwner", Argument: "id", Array: false}
-	ids := []string{"1", "2", "3"}
-	selectionSet := []ast.Selection{
-		&ast.Field{
-			Alias:            "_bramble_id",
-			Name:             "id",
-			Definition:       schema.Types["Owner"].Fields.ForName("id"),
-			ObjectDefinition: schema.Types["Owner"],
-		},
-		&ast.Field{
-			Alias:            "name",
-			Name:             "name",
-			Definition:       schema.Types["Owner"].Fields.ForName("name"),
-			ObjectDefinition: schema.Types["Owner"],
-		},
-	}
-	step := &QueryPlanStep{
-		ServiceURL:     "http://example.com:8080",
-		ServiceName:    "test",
-		ParentType:     "Gizmo",
-		SelectionSet:   selectionSet,
-		InsertionPoint: []string{"gizmos", "owner"},
-		Then:           nil,
-	}
-	expected := []string{`{ _0: getOwner(id: "1") { _bramble_id: id name } _1: getOwner(id: "2") { _bramble_id: id name } _2: getOwner(id: "3") { _bramble_id: id name } }`}
-	ctx := testContextWithoutVariables(nil)
-	docs, err := buildBoundaryQueryDocuments(ctx, schema, step, ids, boundaryField, 10)
-	require.NoError(t, err)
-	require.Equal(t, expected, docs)
-}
-
-func TestBuildBatchedNonArrayBoundaryQueryDocuments(t *testing.T) {
-	ddl := `
-		type Gizmo {
-			id: ID!
-			color: String!
-			owner: Owner
-		}
-
-		type Owner {
-			id: ID!
-			name: String!
-		}
-
-		type Query {
-			gizmos: [Gizmo!]!
-			getOwner(id: ID!): Owner!
-		}
-	`
-	schema := gqlparser.MustLoadSchema(&ast.Source{Name: "fixture", Input: ddl})
-	boundaryField := BoundaryField{Field: "getOwner", Argument: "id", Array: false}
-	ids := []string{"1", "2", "3"}
-	selectionSet := []ast.Selection{
-		&ast.Field{
-			Alias:            "_bramble_id",
-			Name:             "id",
-			Definition:       schema.Types["Owner"].Fields.ForName("id"),
-			ObjectDefinition: schema.Types["Owner"],
-		},
-		&ast.Field{
-			Alias:            "name",
-			Name:             "name",
-			Definition:       schema.Types["Owner"].Fields.ForName("name"),
-			ObjectDefinition: schema.Types["Owner"],
-		},
-	}
-	step := &QueryPlanStep{
-		ServiceURL:     "http://example.com:8080",
-		ServiceName:    "test",
-		ParentType:     "Gizmo",
-		SelectionSet:   selectionSet,
-		InsertionPoint: []string{"gizmos", "owner"},
-		Then:           nil,
-	}
-	expected := []string{`{ _0: getOwner(id: "1") { _bramble_id: id name } _1: getOwner(id: "2") { _bramble_id: id name } }`, `{ _2: getOwner(id: "3") { _bramble_id: id name } }`}
-	ctx := testContextWithoutVariables(nil)
-	docs, err := buildBoundaryQueryDocuments(ctx, schema, step, ids, boundaryField, 2)
-	require.NoError(t, err)
-	require.Equal(t, expected, docs)
-}
-
-func TestMergeExecutionResults(t *testing.T) {
-	t.Run("merges single map", func(t *testing.T) {
-		inputMap := jsonToInterfaceMap(`{
-			"gizmo": {
-				"_bramble_id": "1",
-				"id": "1",
-				"color": "Gizmo A"
-			}
-		}`)
-
-		result := executionResult{
-			ServiceURL:     "http://service-a",
-			InsertionPoint: []string{},
-			Data:           inputMap,
-		}
-
-		mergedMap, err := mergeExecutionResults([]executionResult{result})
-
-		require.NoError(t, err)
-		require.Equal(t, inputMap, mergedMap)
-	})
-
-	t.Run("merges two top level results", func(t *testing.T) {
-		inputMapA := jsonToInterfaceMap(`{
-			"gizmoA": {
-				"id": "1",
-				"color": "Gizmo A"
-			}
-		}`)
-
-		resultA := executionResult{
-			ServiceURL:     "http://service-a",
-			InsertionPoint: []string{},
-			Data:           inputMapA,
-		}
-
-		inputMapB := jsonToInterfaceMap(`{
-			"gizmoB": {
-				"id": "2",
-				"color": "Gizmo B"
-			}
-		}`)
-
-		resultB := executionResult{
-			ServiceURL:     "http://service-b",
-			InsertionPoint: []string{},
-			Data:           inputMapB,
-		}
-
-		mergedMap, err := mergeExecutionResults([]executionResult{resultA, resultB})
-
-		expected := jsonToInterfaceMap(`{
-			"gizmoA": {
-				"id": "1",
-				"color": "Gizmo A"
-			},
-			"gizmoB": {
-				"id": "2",
-				"color": "Gizmo B"
-			}
-		}`)
-
-		require.NoError(t, err)
-		require.Equal(t, expected, mergedMap)
-	})
-
-	t.Run("merges mid level array", func(t *testing.T) {
-		inputMapA := jsonToInterfaceMap(`{
-			"gizmo": {
-				"_bramble_id": "1",
-				"gadgets": [
-					{"_bramble_id": "GADGET1", "owner": { "_bramble_id": "OWNER1" }},
-					{"_bramble_id": "GADGET3", "owner": { "_bramble_id": "OWNER3" }},
-					{"_bramble_id": "GADGET2", "owner": null}
-				]
-			}
-		}`)
-
-		resultA := executionResult{
-			ServiceURL:     "http://service-a",
-			InsertionPoint: []string{},
-			Data:           inputMapA,
-		}
-
-		inputMapB := jsonToInterfaceSlice(`[
-			{
-				"_bramble_id": "OWNER1",
-				"name": "008"
-			}
-		]`)
-
-		resultB := executionResult{
-			ServiceURL:     "http://service-b",
-			InsertionPoint: []string{"gizmo", "gadgets", "owner"},
-			Data:           inputMapB,
-		}
-
-		mergedMap, err := mergeExecutionResults([]executionResult{resultA, resultB})
-
-		expected := jsonToInterfaceMap(`
-		{
-			"gizmo": {
-				"gadgets": [
-					{
-						"_bramble_id": "GADGET1",
-						"owner": {
-							"_bramble_id": "OWNER1",
-							"name": "008"
-						}
-					},
-					{
-						"_bramble_id": "GADGET3",
-						"owner": {
-							"_bramble_id": "OWNER3"
-						}
-					},
-					{
-						"_bramble_id": "GADGET2",
-						"owner": null
-					}
-				],
-				"_bramble_id": "1"
-			}
-		}`)
-
-		require.NoError(t, err)
-		require.Equal(t, expected, mergedMap)
-	})
-
-	t.Run("merges nested mid-level array", func(t *testing.T) {
-		inputMapA := jsonToInterfaceMap(`{
-			"gizmo": {
-				"_bramble_id": "1",
-				"gadgets": [
-					[
-						{"_bramble_id": "GADGET1", "owner": { "_bramble_id": "OWNER1" }},
-						{"_bramble_id": "GADGET3", "owner": { "_bramble_id": "OWNER3" }}
-					],
-					[
-						{"_bramble_id": "GADGET2", "owner": null}
-					]
-				]
-			}
-		}`)
-
-		resultA := executionResult{
-			ServiceURL:     "http://service-a",
-			InsertionPoint: []string{},
-			Data:           inputMapA,
-		}
-
-		inputMapB := jsonToInterfaceSlice(`[
-			{
-				"_bramble_id": "OWNER1",
-				"name": "008"
-			}
-		]`)
-
-		resultB := executionResult{
-			ServiceURL:     "http://service-b",
-			InsertionPoint: []string{"gizmo", "gadgets", "owner"},
-			Data:           inputMapB,
-		}
-
-		mergedMap, err := mergeExecutionResults([]executionResult{resultA, resultB})
-
-		expected := jsonToInterfaceMap(`
-		{
-			"gizmo": {
-				"gadgets": [
-					[
-						{
-							"_bramble_id": "GADGET1",
-							"owner": {
-								"_bramble_id": "OWNER1",
-								"name": "008"
-							}
-						},
-						{
-							"_bramble_id": "GADGET3",
-							"owner": {
-								"_bramble_id": "OWNER3"
-							}
-						}
-					],
-					[
-						{
-							"_bramble_id": "GADGET2",
-							"owner": null
-						}
-					]
-				],
-				"_bramble_id": "1"
-			}
-		}`)
-
-		require.NoError(t, err)
-		require.Equal(t, expected, mergedMap)
-	})
-
-	t.Run("merges root step with child step (root step returns object, boundary field is non array)", func(t *testing.T) {
-		inputMapA := jsonToInterfaceMap(`{
-			"gizmo": {
-				"id": "1",
-				"color": "Gizmo A",
-				"owner": {
-					"_bramble_id": "1"
-				}
-			}
-		}`)
-
-		resultA := executionResult{
-			ServiceURL:     "http://service-a",
-			InsertionPoint: []string{},
-			Data:           inputMapA,
-		}
-
-		inputSliceB := jsonToInterfaceSlice(`[
-			{
-				"_bramble_id": "1",
-				"name": "Owner A"
-			}
-		]`)
-
-		resultB := executionResult{
-			ServiceURL:     "http://service-b",
-			InsertionPoint: []string{"gizmo", "owner"},
-			Data:           inputSliceB,
-		}
-
-		mergedMap, err := mergeExecutionResults([]executionResult{resultA, resultB})
-
-		expected := jsonToInterfaceMap(`{
-			"gizmo": {
-				"id": "1",
-				"color": "Gizmo A",
-				"owner": {
-					"_bramble_id": "1",
-					"name": "Owner A"
-				}
-			}
-		}`)
-
-		require.NoError(t, err)
-		require.Equal(t, expected, mergedMap)
-	})
-
-	t.Run("merges root step with child step (root step returns array, boundary field is non array)", func(t *testing.T) {
-		inputMapA := jsonToInterfaceMap(`{
-			"gizmos": [
-				{
-					"id": "1",
-					"color": "RED",
-					"owner": {
-						"_bramble_id": "4"
-					}
-				},
-				{
-					"id": "2",
-					"color": "GREEN",
-					"owner": {
-						"_bramble_id": "5"
-					}
-				},
-				{
-					"id": "3",
-					"color": "BLUE",
-					"owner": {
-						"_bramble_id": "6"
-					}
-				}
-			]
-		}`)
-
-		resultA := executionResult{
-			ServiceURL:     "http://service-a",
-			InsertionPoint: []string{},
-			Data:           inputMapA,
-		}
-
-		inputSliceB := jsonToInterfaceSlice(`[
-			{
-				"_bramble_id": "4",
-				"name": "Owner A"
-			},
-			{
-				"_bramble_id": "5",
-				"name": "Owner B"
-			},
-			{
-				"_bramble_id": "6",
-				"name": "Owner C"
-			}
-		]`)
-
-		resultB := executionResult{
-			ServiceURL:     "http://service-b",
-			InsertionPoint: []string{"gizmos", "owner"},
-			Data:           inputSliceB,
-		}
-
-		mergedMap, err := mergeExecutionResults([]executionResult{resultA, resultB})
-
-		expected := jsonToInterfaceMap(`{
-			"gizmos": [
-				{
-					"id": "1",
-					"color": "RED",
-					"owner": {
-						"_bramble_id": "4",
-						"name": "Owner A"
-					}
-				},
-				{
-					"id": "2",
-					"color": "GREEN",
-					"owner": {
-						"_bramble_id": "5",
-						"name": "Owner B"
-					}
-				},
-				{
-					"id": "3",
-					"color": "BLUE",
-					"owner": {
-						"_bramble_id": "6",
-						"name": "Owner C"
-					}
-				}
-			]
-		}`)
-
-		require.NoError(t, err)
-		require.Equal(t, expected, mergedMap)
-	})
-
-	t.Run("merges root step with child step (root step returns array, boundary field is array)", func(t *testing.T) {
-		inputMapA := jsonToInterfaceMap(`{
-			"gizmos": [
-				{
-					"id": "1",
-					"color": "RED",
-					"owner": {
-						"_bramble_id": "4"
-					}
-				},
-				{
-					"id": "2",
-					"color": "GREEN",
-					"owner": {
-						"_bramble_id": "5"
-					}
-				},
-				{
-					"id": "3",
-					"color": "BLUE",
-					"owner": {
-						"_bramble_id": "6"
-					}
-				}
-			]
-		}`)
-
-		resultA := executionResult{
-			ServiceURL:     "http://service-a",
-			InsertionPoint: []string{},
-			Data:           inputMapA,
-		}
-
-		inputSliceB := jsonToInterfaceSlice(`[
-			{
-				"_bramble_id": "4",
-				"name": "Owner A"
-			},
-			{
-				"_bramble_id": "5",
-				"name": "Owner B"
-			},
-			{
-				"_bramble_id": "6",
-				"name": "Owner C"
-			}
-		]`)
-
-		resultB := executionResult{
-			ServiceURL:     "http://service-b",
-			InsertionPoint: []string{"gizmos", "owner"},
-			Data:           inputSliceB,
-		}
-
-		mergedMap, err := mergeExecutionResults([]executionResult{resultA, resultB})
-
-		expected := jsonToInterfaceMap(`{
-			"gizmos": [
-				{
-					"id": "1",
-					"color": "RED",
-					"owner": {
-						"_bramble_id": "4",
-						"name": "Owner A"
-					}
-				},
-				{
-					"id": "2",
-					"color": "GREEN",
-					"owner": {
-						"_bramble_id": "5",
-						"name": "Owner B"
-					}
-				},
-				{
-					"id": "3",
-					"color": "BLUE",
-					"owner": {
-						"_bramble_id": "6",
-						"name": "Owner C"
-					}
-				}
-			]
-		}`)
-
-		require.NoError(t, err)
-		require.Equal(t, expected, mergedMap)
-	})
-
-	t.Run("merges using '_bramble_id'", func(t *testing.T) {
-		inputMapA := jsonToInterfaceMap(`{
-			"gizmos": [
-				{
-					"_bramble_id": "1",
-					"color": "RED",
-					"owner": {
-						"_bramble_id": "4"
-					}
-				},
-				{
-					"_bramble_id": "2",
-					"color": "GREEN",
-					"owner": {
-						"_bramble_id": "5"
-					}
-				},
-				{
-					"_bramble_id": "3",
-					"color": "BLUE",
-					"owner": {
-						"_bramble_id": "6"
-					}
-				}
-			]
-		}`)
-
-		resultA := executionResult{
-			ServiceURL:     "http://service-a",
-			InsertionPoint: []string{},
-			Data:           inputMapA,
-		}
-
-		inputSliceB := jsonToInterfaceSlice(`[
-			{
-				"_bramble_id": "4",
-				"name": "Owner A"
-			},
-			{
-				"_bramble_id": "5",
-				"name": "Owner B"
-			},
-			{
-				"_bramble_id": "6",
-				"name": "Owner C"
-			}
-		]`)
-
-		resultB := executionResult{
-			ServiceURL:     "http://service-b",
-			InsertionPoint: []string{"gizmos", "owner"},
-			Data:           inputSliceB,
-		}
-
-		mergedMap, err := mergeExecutionResults([]executionResult{resultA, resultB})
-
-		expected := jsonToInterfaceMap(`{
-			"gizmos": [
-				{
-					"_bramble_id": "1",
-					"color": "RED",
-					"owner": {
-						"_bramble_id": "4",
-						"name": "Owner A"
-					}
-				},
-				{
-					"_bramble_id": "2",
-					"color": "GREEN",
-					"owner": {
-						"_bramble_id": "5",
-						"name": "Owner B"
-					}
-				},
-				{
-					"_bramble_id": "3",
-					"color": "BLUE",
-					"owner": {
-						"_bramble_id": "6",
-						"name": "Owner C"
-					}
-				}
-			]
-		}`)
-
-		require.NoError(t, err)
-		require.Equal(t, expected, mergedMap)
-	})
-
-	t.Run("merges with nil destination", func(t *testing.T) {
-		inputMapA := jsonToInterfaceMap(`{
-			"gizmo": {
-				"_bramble_id": "1",
-				"gadgets": [
-					[
-						{"_bramble_id": "GADGET1", "details": { "owner": {"_bramble_id": "OWNER1" }}}
-					],
-					[
-						{"_bramble_id": "GADGET2", "details": null}
-					]
-				]
-			}
-		}`)
-
-		resultA := executionResult{
-			ServiceURL:     "http://service-a",
-			InsertionPoint: []string{},
-			Data:           inputMapA,
-		}
-
-		inputMapB := jsonToInterfaceSlice(`[
-			{
-				"_bramble_id": "OWNER1",
-				"name": "Alice"
-			}
-		]`)
-
-		resultB := executionResult{
-			ServiceURL:     "http://service-b",
-			InsertionPoint: []string{"gizmo", "gadgets", "details", "owner"},
-			Data:           inputMapB,
-		}
-
-		mergedMap, err := mergeExecutionResults([]executionResult{resultA, resultB})
-
-		expected := jsonToInterfaceMap(`
-		{
-			"gizmo": {
-				"gadgets": [
-					[
-						{
-							"_bramble_id": "GADGET1",
-							"details": {
-								"owner": {
-									"_bramble_id": "OWNER1",
-									"name": "Alice"
-								}
-							}
-						}
-					],
-					[
-						{
-							"_bramble_id": "GADGET2",
-							"details": null
-						}
-					]
-				],
-				"_bramble_id": "1"
-			}
-		}`)
-
-		require.NoError(t, err)
-		require.Equal(t, expected, mergedMap)
-	})
-}
-
-func TestUnionAndTrimSelectionSet(t *testing.T) {
-	schemaString := `
-		directive @boundary on OBJECT
-		interface Tool {
-			id: ID!
-			name: String!
-		}
-
-		union GadgetOrGizmo = Gadget | Gizmo
-
-		type Gizmo @boundary {
-			id: ID!
-		}
-
-		type Gadget @boundary {
-			id: ID!
-		}
-
-		type Agent {
-			id: ID!
-			name: String!
-			country: Country!
-		}
-
-		type Country {
-			id: ID!
-			name: String!
-		}
-
-		type GizmoImplementation implements Tool {
-			id: ID!
-			name: String!
-			gizmos: [Gizmo!]!
-		}
-
-		type GadgetImplementation implements Tool {
-			id: ID!
-			name: String!
-			gadgets: [Gadget!]!
-		}
-
-		type Query {
-			tool(id: ID!): Tool!
-		}`
-
-	schema := gqlparser.MustLoadSchema(&ast.Source{Input: schemaString})
-	ctx := testContextWithoutVariables(nil)
-
-	t.Run("does not touch simple selection sets", func(t *testing.T) {
-		selectionSet := ast.SelectionSet{
-			&ast.Field{
-				Alias:            "id",
-				Name:             "id",
-				Definition:       schema.Types["Agent"].Fields.ForName("id"),
-				ObjectDefinition: schema.Types["Agent"],
-			},
-			&ast.Field{
-				Alias:            "name",
-				Name:             "name",
-				Definition:       schema.Types["Agent"].Fields.ForName("name"),
-				ObjectDefinition: schema.Types["Agent"],
-			},
-			&ast.Field{
-				Alias:            "country",
-				Name:             "country",
-				Definition:       schema.Types["Agent"].Fields.ForName("country"),
-				ObjectDefinition: schema.Types["Agent"],
-				SelectionSet: []ast.Selection{
-					&ast.Field{
-						Alias:            "id",
-						Name:             "id",
-						Definition:       schema.Types["Country"].Fields.ForName("id"),
-						ObjectDefinition: schema.Types["Country"],
-					},
-					&ast.Field{
-						Alias:            "name",
-						Name:             "name",
-						Definition:       schema.Types["Country"].Fields.ForName("name"),
-						ObjectDefinition: schema.Types["Country"],
-					},
-				},
-			},
-		}
-
-		filtered := unionAndTrimSelectionSet("", schema, selectionSet)
-		require.Equal(t, selectionSet, filtered)
-	})
-
-	t.Run("removes duplicate leaf values and merges composite scopes", func(t *testing.T) {
-		selectionSet := ast.SelectionSet{
-			&ast.Field{
-				Alias:            "name",
-				Name:             "name",
-				Definition:       schema.Types["Agent"].Fields.ForName("name"),
-				ObjectDefinition: schema.Types["Agent"],
-			},
-			&ast.Field{
-				Alias:            "name",
-				Name:             "name",
-				Definition:       schema.Types["Agent"].Fields.ForName("name"),
-				ObjectDefinition: schema.Types["Agent"],
-			},
-			&ast.Field{
-				Alias:            "country",
-				Name:             "country",
-				Definition:       schema.Types["Agent"].Fields.ForName("country"),
-				ObjectDefinition: schema.Types["Agent"],
-				SelectionSet: []ast.Selection{
-					&ast.Field{
-						Alias:            "id",
-						Name:             "id",
-						Definition:       schema.Types["Country"].Fields.ForName("id"),
-						ObjectDefinition: schema.Types["Country"],
-					},
-				},
-			},
-			&ast.Field{
-				Alias:            "country",
-				Name:             "country",
-				Definition:       schema.Types["Agent"].Fields.ForName("country"),
-				ObjectDefinition: schema.Types["Agent"],
-				SelectionSet: []ast.Selection{
-					&ast.Field{
-						Alias:            "name",
-						Name:             "name",
-						Definition:       schema.Types["Country"].Fields.ForName("name"),
-						ObjectDefinition: schema.Types["Country"],
-					},
-				},
-			},
-		}
-
-		filtered := unionAndTrimSelectionSet("", schema, selectionSet)
-		require.Equal(t, formatSelectionSetSingleLine(ctx, schema, filtered), "{ name country { id name } }")
-	})
-
-	t.Run("removes field duplicates from inline fragment", func(t *testing.T) {
-		initialSelectionSet := ast.SelectionSet{
-			&ast.Field{
-				Alias:            "id",
-				Name:             "id",
-				Definition:       schema.Types["Tool"].Fields.ForName("id"),
-				ObjectDefinition: schema.Types["Tool"],
-			},
-			&ast.Field{
-				Alias:            "name",
-				Name:             "name",
-				Definition:       schema.Types["Tool"].Fields.ForName("name"),
-				ObjectDefinition: schema.Types["Tool"],
-			},
-			&ast.InlineFragment{
-				TypeCondition: schema.Types["GizmoImplementation"].Name,
-				SelectionSet: []ast.Selection{
-					&ast.Field{
-						Alias:            "id",
-						Name:             "id",
-						Definition:       schema.Types["GizmoImplementation"].Fields.ForName("id"),
-						ObjectDefinition: schema.Types["GizmoImplementation"],
-					},
-					&ast.Field{
-						Alias:            "name",
-						Name:             "name",
-						Definition:       schema.Types["GizmoImplementation"].Fields.ForName("name"),
-						ObjectDefinition: schema.Types["GizmoImplementation"],
-					},
-					&ast.Field{
-						Alias:            "gizmos",
-						Name:             "gizmos",
-						Definition:       schema.Types["GizmoImplementation"].Fields.ForName("gizmos"),
-						ObjectDefinition: schema.Types["GizmoImplementation"],
-						SelectionSet: []ast.Selection{
-							&ast.Field{
-								Alias:            "id",
-								Name:             "id",
-								Definition:       schema.Types["Gizmo"].Fields.ForName("id"),
-								ObjectDefinition: schema.Types["Gizmo"],
-							},
-						},
-					},
-				},
-				ObjectDefinition: schema.Types["GizmoImplementation"],
-			},
-		}
-
-		expected := ast.SelectionSet{
-			&ast.Field{
-				Alias:            "id",
-				Name:             "id",
-				Definition:       schema.Types["Tool"].Fields.ForName("id"),
-				ObjectDefinition: schema.Types["Tool"],
-			},
-			&ast.Field{
-				Alias:            "name",
-				Name:             "name",
-				Definition:       schema.Types["Tool"].Fields.ForName("name"),
-				ObjectDefinition: schema.Types["Tool"],
-			},
-			&ast.InlineFragment{
-				TypeCondition: schema.Types["GizmoImplementation"].Name,
-				SelectionSet: []ast.Selection{
-					&ast.Field{
-						Alias:            "gizmos",
-						Name:             "gizmos",
-						Definition:       schema.Types["GizmoImplementation"].Fields.ForName("gizmos"),
-						ObjectDefinition: schema.Types["GizmoImplementation"],
-						SelectionSet: []ast.Selection{
-							&ast.Field{
-								Alias:            "id",
-								Name:             "id",
-								Definition:       schema.Types["Gizmo"].Fields.ForName("id"),
-								ObjectDefinition: schema.Types["Gizmo"],
-							},
-						},
-					},
-				},
-				ObjectDefinition: schema.Types["GizmoImplementation"],
-			},
-		}
-
-		filtered := unionAndTrimSelectionSet("GizmoImplementation", schema, initialSelectionSet)
-		require.Equal(t, formatSelectionSetSingleLine(ctx, schema, expected), formatSelectionSetSingleLine(ctx, schema, filtered))
-	})
-
-	t.Run("removes inline fragment if it only contains duplicate selections", func(t *testing.T) {
-		initialSelectionSet := ast.SelectionSet{
-			&ast.Field{
-				Alias:            "id",
-				Name:             "id",
-				Definition:       schema.Types["Tool"].Fields.ForName("id"),
-				ObjectDefinition: schema.Types["Tool"],
-			},
-			&ast.Field{
-				Alias:            "name",
-				Name:             "name",
-				Definition:       schema.Types["Tool"].Fields.ForName("name"),
-				ObjectDefinition: schema.Types["Tool"],
-			},
-			&ast.InlineFragment{
-				TypeCondition: schema.Types["GizmoImplementation"].Name,
-				SelectionSet: []ast.Selection{
-					&ast.Field{
-						Alias:            "id",
-						Name:             "id",
-						Definition:       schema.Types["GizmoImplementation"].Fields.ForName("id"),
-						ObjectDefinition: schema.Types["GizmoImplementation"],
-					},
-					&ast.Field{
-						Alias:            "name",
-						Name:             "name",
-						Definition:       schema.Types["GizmoImplementation"].Fields.ForName("name"),
-						ObjectDefinition: schema.Types["GizmoImplementation"],
-					},
-				},
-				ObjectDefinition: schema.Types["GizmoImplementation"],
-			},
-		}
-
-		expected := ast.SelectionSet{
-			&ast.Field{
-				Alias:            "id",
-				Name:             "id",
-				Definition:       schema.Types["Tool"].Fields.ForName("id"),
-				ObjectDefinition: schema.Types["Tool"],
-			},
-			&ast.Field{
-				Alias:            "name",
-				Name:             "name",
-				Definition:       schema.Types["Tool"].Fields.ForName("name"),
-				ObjectDefinition: schema.Types["Tool"],
-			},
-		}
-
-		filtered := unionAndTrimSelectionSet("GizmoImplementation", schema, initialSelectionSet)
-		require.Equal(t, formatSelectionSetSingleLine(ctx, schema, expected), formatSelectionSetSingleLine(ctx, schema, filtered))
-	})
-
-	t.Run("removes inline fragment that does not match typename", func(t *testing.T) {
-		initialSelectionSet := ast.SelectionSet{
-			&ast.Field{
-				Alias:            "id",
-				Name:             "id",
-				Definition:       schema.Types["Tool"].Fields.ForName("id"),
-				ObjectDefinition: schema.Types["Tool"],
-			},
-			&ast.InlineFragment{
-				TypeCondition: schema.Types["GizmoImplementation"].Name,
-				SelectionSet: []ast.Selection{
-					&ast.Field{
-						Alias:            "id",
-						Name:             "id",
-						Definition:       schema.Types["GizmoImplementation"].Fields.ForName("id"),
-						ObjectDefinition: schema.Types["GizmoImplementation"],
-					},
-					&ast.Field{
-						Alias:            "name",
-						Name:             "name",
-						Definition:       schema.Types["GizmoImplementation"].Fields.ForName("name"),
-						ObjectDefinition: schema.Types["GizmoImplementation"],
-					},
-				},
-				ObjectDefinition: schema.Types["GizmoImplementation"],
-			},
-			&ast.InlineFragment{
-				TypeCondition: schema.Types["GadgetImplementation"].Name,
-				SelectionSet: []ast.Selection{
-					&ast.Field{
-						Alias:            "name",
-						Name:             "name",
-						Definition:       schema.Types["GadgetImplementation"].Fields.ForName("name"),
-						ObjectDefinition: schema.Types["GadgetImplementation"],
-					},
-				},
-				ObjectDefinition: schema.Types["GadgetImplementation"],
-			},
-		}
-
-		expected := ast.SelectionSet{
-			&ast.Field{
-				Alias:            "id",
-				Name:             "id",
-				Definition:       schema.Types["Tool"].Fields.ForName("id"),
-				ObjectDefinition: schema.Types["Tool"],
-			},
-			&ast.InlineFragment{
-				TypeCondition: schema.Types["GizmoImplementation"].Name,
-				SelectionSet: []ast.Selection{
-					&ast.Field{
-						Alias:            "name",
-						Name:             "name",
-						Definition:       schema.Types["GizmoImplementation"].Fields.ForName("name"),
-						ObjectDefinition: schema.Types["GizmoImplementation"],
-					},
-				},
-				ObjectDefinition: schema.Types["GizmoImplementation"],
-			},
-		}
-
-		filtered := unionAndTrimSelectionSet("GizmoImplementation", schema, initialSelectionSet)
-		require.Equal(t, formatSelectionSetSingleLine(ctx, schema, expected), formatSelectionSetSingleLine(ctx, schema, filtered))
-	})
-
-	t.Run("works with unions", func(t *testing.T) {
-		initialSelectionSet := ast.SelectionSet{
-			&ast.InlineFragment{
-				TypeCondition: schema.Types["Gizmo"].Name,
-				SelectionSet: []ast.Selection{
-					&ast.Field{
-						Alias:            "id",
-						Name:             "id",
-						Definition:       schema.Types["Gizmo"].Fields.ForName("id"),
-						ObjectDefinition: schema.Types["Gizmo"],
-					},
-				},
-				ObjectDefinition: schema.Types["GadgetOrGizmo"],
-			},
-			&ast.InlineFragment{
-				TypeCondition: schema.Types["Gadget"].Name,
-				SelectionSet: []ast.Selection{
-					&ast.Field{
-						Alias:            "name",
-						Name:             "name",
-						Definition:       schema.Types["Gadget"].Fields.ForName("name"),
-						ObjectDefinition: schema.Types["Gadget"],
-					},
-				},
-				ObjectDefinition: schema.Types["GadgetOrGizmo"],
-			},
-		}
-
-		expected := ast.SelectionSet{
-			&ast.InlineFragment{
-				TypeCondition: schema.Types["Gadget"].Name,
-				SelectionSet: []ast.Selection{
-					&ast.Field{
-						Alias:            "name",
-						Name:             "name",
-						Definition:       schema.Types["Gadget"].Fields.ForName("name"),
-						ObjectDefinition: schema.Types["Gadget"],
-					},
-				},
-				ObjectDefinition: schema.Types["GadgetOrGizmo"],
-			},
-		}
-
-		filtered := unionAndTrimSelectionSet("Gadget", schema, initialSelectionSet)
-		require.Equal(t, formatSelectionSetSingleLine(ctx, schema, expected), formatSelectionSetSingleLine(ctx, schema, filtered))
-	})
 }
 
 func TestBubbleUpNullValuesInPlace(t *testing.T) {
